refactor: split secret scan detectors
