+++ src/security/secret-scan/detectors/format.ts
@@ -1,0 +1,132 @@
+import type { RegexDetector } from "./types.js";
+
+export const formatDetectors: RegexDetector[] = [
+  {
+    id: "pem-private-key",
+    kind: "format",
+    confidence: "high",
+    pattern: String.raw`-----BEGIN [A-Z ]*PRIVATE KEY-----[\s\S]+?-----END [A-Z ]*PRIVATE KEY-----`,
+    redact: "full",
+  },
+  {
+    id: "authorization-bearer",
+    kind: "format",
+    confidence: "high",
+    pattern: String.raw`Authorization\s*[:=]\s*Bearer\s+([A-Za-z0-9._\-+=]+)`,
+    flags: "gi",
+    group: 1,
+    redact: "group",
+  },
+  {
+    id: "bearer-inline",
+    kind: "format",
+    confidence: "medium",
+    pattern: String.raw`\bBearer\s+([A-Za-z0-9._\-+=]{18,})\b`,
+    flags: "gi",
+    group: 1,
+    redact: "group",
+  },
+  {
+    id: "token-sk",
+    kind: "format",
+    confidence: "high",
+    pattern: String.raw`\b(sk-[A-Za-z0-9_-]{8,})\b`,
+    group: 1,
+    redact: "group",
+  },
+  {
+    id: "token-ghp",
+    kind: "format",
+    confidence: "high",
+    pattern: String.raw`\b(ghp_[A-Za-z0-9]{20,})\b`,
+    group: 1,
+    redact: "group",
+  },
+  {
+    id: "token-github-pat",
+    kind: "format",
+    confidence: "high",
+    pattern: String.raw`\b(github_pat_[A-Za-z0-9_]{20,})\b`,
+    group: 1,
+    redact: "group",
+  },
+  {
+    id: "token-slack",
+    kind: "format",
+    confidence: "high",
+    pattern: String.raw`\b(xox(?:[baprs]|o)-[A-Za-z0-9-]{10,})\b`,
+    group: 1,
+    redact: "group",
+  },
+  {
+    id: "token-slack-webhook",
+    kind: "format",
+    confidence: "high",
+    pattern: String.raw`https://hooks\.slack\.com/services/[A-Za-z0-9]+/[A-Za-z0-9]+/[A-Za-z0-9]+`,
+    redact: "full",
+  },
+  {
+    id: "token-slack-app",
+    kind: "format",
+    confidence: "high",
+    pattern: String.raw`\b(xapp-[A-Za-z0-9-]{10,})\b`,
+    group: 1,
+    redact: "group",
+  },
+  {
+    id: "token-google-ai",
+    kind: "format",
+    confidence: "high",
+    pattern: String.raw`\b(gsk_[A-Za-z0-9_-]{10,})\b`,
+    group: 1,
+    redact: "group",
+  },
+  {
+    id: "token-google-api",
+    kind: "format",
+    confidence: "high",
+    pattern: String.raw`\b(AIza[0-9A-Za-z\-_]{20,})\b`,
+    group: 1,
+    redact: "group",
+  },
+  {
+    id: "token-perplexity",
+    kind: "format",
+    confidence: "high",
+    pattern: String.raw`\b(pplx-[A-Za-z0-9_-]{10,})\b`,
+    group: 1,
+    redact: "group",
+  },
+  {
+    id: "token-npm",
+    kind: "format",
+    confidence: "high",
+    pattern: String.raw`\b(npm_[A-Za-z0-9]{10,})\b`,
+    group: 1,
+    redact: "group",
+  },
+  {
+    id: "token-telegram",
+    kind: "format",
+    confidence: "high",
+    pattern: String.raw`\b(\d{6,}:[A-Za-z0-9_-]{20,})\b`,
+    group: 1,
+    redact: "group",
+  },
+  {
+    id: "token-aws-access",
+    kind: "format",
+    confidence: "high",
+    pattern: String.raw`\b(A(?:KI|SI)A[0-9A-Z]{16})\b`,
+    group: 1,
+    redact: "group",
+  },
+  {
+    id: "jwt",
+    kind: "format",
+    confidence: "medium",
+    pattern: String.raw`\b([A-Za-z0-9_-]{10,}\.[A-Za-z0-9_-]{10,}\.[A-Za-z0-9_-]{10,})\b`,
+    group: 1,
+    redact: "group",
+  },
+];
