+++ src/security/secret-scan/detectors/index.ts
@@ -1,0 +1,3 @@
+export { addEntropyDetections } from "./entropy.js";
+export { addRegexDetections } from "./regex.js";
+export type { Redaction } from "./types.js";
